chore: moves secret generation to its own package
